Add a version string.
It will be embeded into libWireCellUtil and is available via CLI and
the Main app.  The version string is set at build time based on the
state of the git source repo.  This won't work from a source tar ball
and we can cross that bridge if we come to it.

Fix #112.