Maybe for real fix the out-of-order problem.
Despite what TBB docs say, it seems function nodes do NOT have a FIFO
input buffer.  Bolting on a sequencer node after each node that needs
it seems to give desired behavior.  This requires elevating the TBB
level message to be a pair of a sequence number and the payload "any".