Add ability to select traces by plane.
There is a stand-alone PlaneSelector to do this and DNNROIFinding
changes from cbeg/cend to using the same pattern internally.

This is based on a channel selection from OmnibusSigProc that is moved
into a new Aux/PlaneTools.h.  But, OmnibusSigProc still is left as-is.
It should be factored to use this moved code.